# Optimize fused allreduce in raw program