<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
Performance optimization

### PR changes
Others

### Describe
1. Increase the overlap of the calc stream and comm stream after fusing the allreduce vars.
2. Move the `coalesce_tensor` before grad generation, so that the `copy_data` is not needed.

#### Test
GPT model with dp=8

||Without Fuse|With Fuse|Gain|
|:--:|:--:|:--:|:--:|
|FP32| 68889 tokens/s |70822 tokens/s| +2.3%|
|AMP|123968 tokens/s | 133408 tokens/s | +7.6% |

**FP32 Loss curve**
<img width="350" alt="Screen Shot 2021-08-03 at 6 46 32 PM" src="https://user-images.githubusercontent.com/25279174/128003164-b7c4cee2-8d96-4c64-9a60-be20021e99e4.png">
Max Diff: 0.038

**FP16 Loss curve**
<img width="350" alt="Screen Shot 2021-08-03 at 7 05 15 PM" src="https://user-images.githubusercontent.com/25279174/128005289-f32c479f-4c0b-4e2d-a0cc-bf5ac0b479a7.png">
Max Diff: 0.073